requirements.txt
asgiref==3.7.2
contourpy==1.1.0
cycler==0.11.0
Django==4.2.3
django-appconf==1.0.5
django-select2==8.1.2
django-smart-selects==1.6.0
fonttools==4.41.1
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly==5.15.0
psycopg2-binary==2.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.4
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
tzdata==2023.3
